Extract the interface of MusicFolderService
 - MusicFolderService concrete class has been moved to the scanner package
 - Extract methods that do not write to the MusicFolder table into the
   MusicFolderService interface
